fix #2446, pc0Flow for all boilers, fix #2411

+++ src/devices/boiler.h
@@ -194,7 +194,7 @@ class Boiler : public EMSdevice {
     uint16_t maintenanceTime_;
 
     // heatpump
-    uint8_t  hpPower_;
+    uint16_t hpPower_;
     uint8_t  hpCompOn_;
     uint8_t  hpBrinePumpSpd_;
     uint8_t  hpCompSpd_;
